dependency tree: entities creation & registration

+++ core/component/position.go
@@ -1,7 +1,7 @@
 package component
 
 type Position struct {
-	X, Y float64
+	X, Y int32
 }
 
 func (p *Position) IsComponent() bool {
